# src/main/resources/api.yaml
swagger: "2.0"
info:
  description: "Swagger for Budget App"
  version: "1.0.0"
  title: "Swagger Petstore"
tags:
  - name: "income"
    description: "Income"
  - name: "group"
    description: "Group"
schemes:
  - "http"
paths:
  /income:
    post:
      tags:
        - "income"
      summary: "Add a new income"
      description: ""
      operationId: "addIncome"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Income object that needs to be added"
          required: true
          schema:
            $ref: "#/definitions/Income"
      responses:
        "201" :
          description: "Created"
          schema:
            $ref: "#/definitions/Income"
        "405":
          description: "Invalid input"
          schema:
            $ref: '#/definitions/ModelApiResponse'
    put:
      tags:
        - "income"
      summary: "Update income"
      description: ""
      operationId: "updateIncome"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Income object that needs to be added"
          required: true
          schema:
            $ref: "#/definitions/Income"
      responses:
        "200":
          description: "Updated"
          schema:
            $ref: "#/definitions/Income"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Pet not found"
        "405":
          description: "Validation exception"
    get:
      tags:
        - "income"
      summary: "Get all Income"
      description: ""
      operationId: "getAllIncome"
      produces:
        - "application/json"
      responses:
        "200":
          description: "All Income"
          schema:
            type: array
            items:
              $ref: "#/definitions/Income"
  /income/{incomeId}:
    get:
      tags:
        - "income"
      summary: "Find income by ID"
      description: "Returns a single income"
      operationId: "getIncomeById"
      produces:
        - "application/json"
      parameters:
        - name: "incomeId"
          in: "path"
          description: "ID of income to return"
          required: true
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Income"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Income not found"
    delete:
      tags:
        - "income"
      summary: "Delete income by ID"
      description: "Deletes a single income"
      operationId: "deleteIncomeById"
      parameters:
        - name: "incomeId"
          in: "path"
          description: "ID of income to delete"
          required: true
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Income not found"
  /group:
    post:
      tags:
        - "group"
      summary: "Add a new group"
      description: ""
      operationId: "addGroup"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Group object that needs to be added"
          required: true
          schema:
            $ref: "#/definitions/Group"
      responses:
        "201":
          description: "Created"
          schema:
            $ref: "#/definitions/Group"
        "405":
          description: "Invalid input"
          schema:
            $ref: '#/definitions/ModelApiResponse'
    put:
      tags:
        - "group"
      summary: "Update group"
      description: ""
      operationId: "updateGroup"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Group object that needs to be added"
          required: true
          schema:
            $ref: "#/definitions/Group"
      responses:
        "200":
          description: "Updated"
          schema:
            $ref: "#/definitions/Group"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Pet not found"
        "405":
          description: "Validation exception"
    get:
      tags:
        - "group"
      summary: "Get all Group"
      description: ""
      operationId: "getAllGroup"
      produces:
        - "application/json"
      responses:
        "200":
          description: "All Group"
          schema:
            type: array
            items:
              $ref: "#/definitions/Group"
  /group/{groupId}:
    get:
      tags:
        - "group"
      summary: "Find group by ID"
      description: "Returns a single group"
      operationId: "getGroupById"
      produces:
        - "application/json"
      parameters:
        - name: "groupId"
          in: "path"
          description: "ID of group to return"
          required: true
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Group"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Group not found"
    delete:
      tags:
        - "group"
      summary: "Delete group by ID"
      description: "Deletes a single group"
      operationId: "deleteGroupById"
      parameters:
        - name: "groupId"
          in: "path"
          description: "ID of group to delete"
          required: true
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Group not found"
definitions:
  Income:
    type: "object"
    required:
      - "amount"
      - "description"
    properties:
      id:
        type: "string"
        format: "uuid"
      amount:
        type: "integer"
        format: "int64"
      description:
        type: "string"
      budget:
        type: "integer"
        format: "int64"
    xml:
      name: "Income"
  Group:
    type: "object"
    required:
      - "name"
      - "description"
    properties:
      id:
        type: "string"
        format: "uuid"
      name:
        type: "string"
      description:
        type: "string"
      budget:
        type: "integer"
        format: "int64"
    xml:
      name: "Income"
  ModelApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
externalDocs:
  description: "Find out more about Budget App"
  url: "http://github.com"